Made changes to the DisplayAllUsers comp, app is now complete.
Chatted with Adam, stopped changing the state to hide the games played (cos it's stupid), instead there's a new boolean variable in the state which toggles when the button is pressed. DisplayAllUsers has been tweaked to use that boolean to decide whether or not to display the number of games.